+++ apps/meteor/client/lib/cachedCollections/index.ts
@@ -1,2 +1,3 @@
 export { PrivateCachedCollection, PublicCachedCollection } from './CachedCollection';
 export { CachedCollectionManager } from './CachedCollectionManager';
+export { pipe } from './pipe';