+++ apps/meteor/client/startup/roles.ts
@@ -10,9 +10,7 @@ Meteor.startup(() => {
 	onLoggedIn(async () => {
 		const { roles } = await sdk.rest.get('/v1/roles.list');
 		// if a role is checked before this collection is populated, it will return undefined
-		Roles.store.replaceAll(roles);
-
-		Roles.ready.set(true);
+		Roles.state.replaceAll(roles);
 	});
 
 	type ClientAction = 'inserted' | 'updated' | 'removed' | 'changed';
@@ -20,10 +18,10 @@ Meteor.startup(() => {
 	const events: Record<string, ((role: IRole & { type?: ClientAction }) => void) | undefined> = {
 		changed: (role) => {
 			delete role.type;
-			Roles.upsert({ _id: role._id }, role);
+			Roles.state.store(role);
 		},
 		removed: (role) => {
-			Roles.remove({ _id: role._id });
+			Roles.state.delete(role._id);
 		},
 	};