+++ apps/meteor/client/views/account/security/BackupCodesModal.tsx
@@ -1,9 +1,8 @@
 import { Box, CodeSnippet } from '@rocket.chat/fuselage';
 import { useClipboard } from '@rocket.chat/fuselage-hooks';
+import { GenericModal } from '@rocket.chat/ui-client';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-
-import GenericModal from '../../../components/GenericModal';
 
 type BackupCodesModalProps = {
 	codes: string[];