+++ apps/meteor/client/views/omnichannel/departments/DepartmentsTable/RemoveDepartmentModal.tsx
@@ -1,11 +1,10 @@
 import { Box, Input } from '@rocket.chat/fuselage';
 import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
+import { GenericModal } from '@rocket.chat/ui-client';
 import { useToastMessageDispatch, useEndpoint } from '@rocket.chat/ui-contexts';
 import type { ChangeEvent, FormEvent, ReactElement } from 'react';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-
-import GenericModal from '../../../../components/GenericModal';
 
 type RemoveDepartmentModalProps = {
 	_id: string;