+++ packages/apps-engine/src/definition/accessors/IEmailCreator.ts
@@ -1,10 +1,10 @@
 import type { IEmail } from '../email';
 
 export interface IEmailCreator {
-    /**
-     * Sends an email through Rocket.Chat
-     *
-     * @param email the email data
-     */
-    send(email: IEmail): Promise<void>;
+	/**
+	 * Sends an email through Rocket.Chat
+	 *
+	 * @param email the email data
+	 */
+	send(email: IEmail): Promise<void>;
 }